verilog: LMS almost complete
Implemented the QOA's LMS predictor in verilog.
Seems to almost be done, passes the trivial unit test from a few
commits ago.

I tried to make the pure python implementation use it, but it errored
out after 9000,000 samples. This is probably a good sign, only a couple
of bugs remaining.